oroshine_app/requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.11.0
beautifulsoup4==4.14.3
billiard==4.2.2
bleach==6.3.0
brotli==1.2.0
build==1.3.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
commonmark==0.9.1
coverage==7.13.4
crispy-bootstrap5==2025.6
cron_descriptor==2.0.6
cryptography==45.0.7
defusedxml==0.7.1
Django==5.1.15
django-allauth==65.1.0
django-appconf==1.2.0
django-bootstrap5==25.3
django-celery-beat==2.8.1
django-compression-middleware==0.5.0
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-debug-toolbar==6.1.0
django-extensions==4.1
django-imagekit==6.0.0
django-minify-html==1.14.0
django-phonenumbers==1.0.1
django-picklefield==3.3
django-prometheus==2.4.1
django-q2==1.8.0
django-redis==6.0.0
django-schema-viewer==0.5.3
django-tempus-dominus==5.1.2.17
django-timezone-field==7.1
django_celery_results==2.6.0
django_compressor==4.6.0
dotenv==0.9.9
exceptiongroup==1.3.1
flower==2.0.1
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
graphviz==0.21
gunicorn==23.0.0
hiredis==3.3.0
httplib2==0.31.2
humanize==4.15.0
icalendar==5.0.11
idna==3.11
iniconfig==2.3.0
kombu==5.5.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
marshmallow-dataclass==8.5.14
mdurl==0.1.2
minify_html==0.18.1
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
phonenumbers==9.0.20
pilkit==3.0
pillow==11.3.0
pip-tools==7.5.2
platformdirs==4.5.1
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.4
psycopg==3.1.16
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pygitguardian==1.27.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.2
pyproject_hooks==1.2.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-ulid==3.1.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.3
rcssmin==1.2.2
redis==7.1.0
refreshcss==0.5.1
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rjsmin==1.2.5
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
social-auth-core==4.7.0
soupsieve==2.8
sqlparse==0.5.4
tornado==6.5.4
truststore==0.10.4
typeguard==4.4.4
types-python-dateutil==2.9.0.20251115
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
Werkzeug==3.1.4
wheel==0.45.1
whitenoise==6.10.0
zstandard==0.25.0
